fix vehicle

+++ Racing.Model/Vehicle/VehicleDetail.cs
@@ -14,7 +14,7 @@ namespace Racing.Model.Vehicle
         public string Name { get; set; }
         public string Class { get; set; }
         public string Series { get; set; }
-        public int HP { get; set; }
+        public double HP { get; set; }
         public decimal Weight { get; set; }
         [Display(Name = "Created")]
         public DateTimeOffset CreatedUtc { get; set; }
